refactor(events): move blacklist query helper to utils to honor model→service rule (#411)
